requirements.txt
Django==1.5.1
django-bootstrap-toolkit==2.12.0
django-form-utils
lxml
django-picklefield==0.3.0
django-jsonify